requirements.txt
appnope==0.1.0
backcall==0.2.0
click==7.1.2
decorator==4.4.2
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.1
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pycryptodomex==3.9.8
Pygments==2.6.1
six==1.15.0
traitlets==4.3.3
wcwidth==0.2.5
